Перенес вьюшки и urls для Личного кабинета в Authapp Переименовал mainapp/base.html в mainapp/mainapp-base.html чтобы из Authapp можно было extend шаблон из mainapp Подправил шаблон account и authapp/base.html вместе с контроллером
