+++ TeamHabr/TeamHabr/settings.py
@@ -82,28 +82,28 @@ WSGI_APPLICATION = 'TeamHabr.wsgi.application'
 """
 ТЕКСТ
 """
-# DATABASES = {
-#     'default': {
-#         'ENGINE': 'django.db.backends.sqlite3',
-#         'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
-#     }
-# }
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
+    }
+}
 """
 ТЕКСТ
 """
 
-DATABASES = {
-   'default': {
-
-       'ENGINE': 'django.db.backends.postgresql',
-       'NAME': 'postgres',
-       'USER': 'postgres',
-       'PASSWORD': '2Be-U>Zq2h[!',
-       'HOST': 'teamhabrdb.cjse1wsktjbz.ca-central-1.rds.amazonaws.com',
-       'PORT': '5432',
-
-    }
-}
+# DATABASES = {
+#    'default': {
+#
+#        'ENGINE': 'django.db.backends.postgresql',
+#        'NAME': 'postgres',
+#        'USER': 'postgres',
+#        'PASSWORD': '2Be-U>Zq2h[!',
+#        'HOST': 'teamhabrdb.cjse1wsktjbz.ca-central-1.rds.amazonaws.com',
+#        'PORT': '5432',
+#
+#     }
+# }
 
 
 # Password validation
