+++ TeamHabr/authapp/templates/authapp/account.html
@@ -1,29 +1,34 @@
-{% extends 'authapp/base.html' %}
+{% extends 'mainapp-base.html' %}
 {% load static %}
 
+{% block users %}
+{% endblock %}
+
 {% block content %}
-<a href="{% url 'main:article-create' %}">Создать новую статью</a>
-<div class="user-inf">
-    {% if user.avatar %}
-    <img src="{{ user.avatar }}" alt="" class="avatar" width="40px" height="40px">
-    {% else %}
-    <img src="https://cdn4.iconfinder.com/data/icons/avatars-xmas-giveaway/128/batman_hero_avatar_comics-512.png" alt="" class="avatar" width="30px" height="30px">
-    {% endif %}
-    <p class="user-name">{{ user.name }} {{ user.surname }}</p>
-    <p>Возраст: {{ user.age }}</p>
-    <p>Биография: {{ user.desc }}</p>
-    <div class="btn"><a href="">Редактировать</a></div>
-</div>
+    <div class="user-inf">
+        {% if user.avatar %}
+            <img src="{{ user.avatar }}" alt="" class="avatar" width="40px" height="40px">
+        {% else %}
+            <img src="https://cdn4.iconfinder.com/data/icons/avatars-xmas-giveaway/128/batman_hero_avatar_comics-512.png" alt="" class="avatar" width="30px" height="30px">
+        {% endif %}
+            <p class="user-name">{{ user.name }} {{ user.surname }}</p>
+            <p>Возраст: {{ user.age }}</p>
+            <p>Биография: {{ user.desc }}</p>
 
-<hr>
+        <div class="btn">
+            <a href="#">Редактировать</a>
+        </div>
+    </div>
 
-<div class="btn"><a href="">Создать статью</a></div>
-<div class="habs">
-  <h3 class="title" style='text-align: center'>Мои статьи</h3>
-  {% if user.article %}
-  <p class="article">{{ user.article }}</p>
-  {% else %}
-  <h4 style='text-align: center'>У вас ещё нет статей</h4>
-  {% endif %}
-</div>
+    <hr>
+
+    <a href="{% url 'main:article-create' %}">Создать новую статью</a>
+    <div class="habs">
+      <h3 class="title" style='text-align: center'>Мои статьи</h3>
+      {% if user.article %}
+        <p class="article">{{ user.article }}</p>
+      {% else %}
+        <h4 style='text-align: center'>У вас ещё нет статей</h4>
+      {% endif %}
+    </div>
 {% endblock %}
