+++ TeamHabr/authapp/templates/authapp/base.html
@@ -29,7 +29,7 @@
     <div class="container">
         <div class="login">
             <div class="h2 text-center head">
-                Аутентификация
+                {{ title }}
             </div>
 
             {% block content %}
