+++ TeamHabr/mainapp/templates/mainapp/index.html
@@ -1,5 +1,5 @@
 
-{% extends 'base.html' %}
+{% extends 'mainapp-base.html' %}
 {% load static %}
 {% block content %}
