+++ TeamHabr/mainapp/templates/mainapp/web-dev.html
@@ -1,2 +1,2 @@
-{% extends 'base.html' %}
+{% extends 'mainapp-base.html' %}
 {% load static %}
